requirements.txt
# Core (pin versions in production as needed)
torch>=2.0
transformers>=4.40
peft>=0.11
datasets>=2.16
accelerate>=0.27
pillow>=10.0
tqdm>=4.66
numpy>=1.24
wandb>=0.16
nltk>=3.8
